still working on placeinfo

+++ src/components/Logout/index.js
@@ -1,45 +1,45 @@
-import React from 'react';
-import Swal from 'sweetalert2';
+import React from 'react'
+import { Redirect } from 'react-router-dom'
+import Swal from 'sweetalert2'
 
-const Logout = ({ setIsAuthenticated }) => {
+function Logout() {
   const handleLogout = () => {
 
-    // empty token
-    localStorage.setItem("token", "")
-
-    // success message
+    // Success message
     Swal.fire({
       icon: 'question',
       title: 'Logging Out',
       text: 'Are you sure you want to log out?',
       showCancelButton: true,
-      confirmButtonText: 'Yes',
-    }).then(result => {
+      confirmButtonText: 'Yes'
+    }).then((result) => {
+      // Empty token
+      localStorage.setItem('token', '')
       if (result.value) {
         Swal.fire({
           timer: 1500,
           showConfirmButton: false,
           willOpen: () => {
-            Swal.showLoading();
+            Swal.showLoading()
           },
           willClose: () => {
-            localStorage.setItem('is_authenticated', false);
-            setIsAuthenticated(false);
-          },
-        });
+          }
+        })
+        window.location.reload(false);
+        // return <Redirect to="/" />
       }
-    });
-  };
+    })
+  }
 
   return (
     <button
-      style={{ marginLeft: '12px' }}
       className="muted-button"
       onClick={handleLogout}
+      style={{ marginLeft: '12px' }}
     >
       Logout
     </button>
-  );
-};
+  )
+}
 
-export default Logout;
+export default Logout
